factor out code into BRUSH, CANVAS, and UNDO
