+++ src/VARIABLES.cpp
@@ -42,78 +42,6 @@ int WINDOW_H;
 int DISPLAY_W;
 int DISPLAY_H;
 
-// CANVAS
-bool CANVAS_UPDATE = false;
-float CANVAS_ZOOM = 1.0;
-float CANVAS_X = 0.0;
-float CANVAS_Y = 0.0;
-uint16_t CANVAS_W = 512;
-uint16_t CANVAS_H = 256;
-int16_t CANVAS_MOUSE_X = 0;
-int16_t CANVAS_MOUSE_Y = 0;
-int16_t CANVAS_MOUSE_CLICKX = 0;
-int16_t CANVAS_MOUSE_CLICKY = 0;
-int16_t CANVAS_MOUSE_CLICKPREVX = 0;
-int16_t CANVAS_MOUSE_CLICKPREVY = 0;
-int16_t CANVAS_MOUSE_PREVX = 0;
-int16_t CANVAS_MOUSE_PREVY = 0;
-int16_t CANVAS_MOUSE_CELL_X = 0;
-int16_t CANVAS_MOUSE_CELL_Y = 0;
-int16_t CANVAS_PITCH = (sizeof(COLOR) * CANVAS_W);
-
-SDL_Texture* BG_GRID_TEXTURE;
-int16_t CELL_W = 16;
-int16_t CELL_H = 16;
-int16_t BG_GRID_W = 0;
-int16_t BG_GRID_H = 0;
-
-float CANVAS_PANX = 0.0;
-float CANVAS_PANY = 0.0;
-
-float CANVAS_X_ANIM = 0.0;
-float CANVAS_Y_ANIM = 0.0;
-float CANVAS_W_ANIM = 0.0;
-float CANVAS_H_ANIM = 0.0;
-float CELL_W_ANIM = 0.0;
-float CELL_H_ANIM = 0.0;
-
-// LAYER
-uint16_t CURRENT_LAYER = 0;
-COLOR* CURRENT_LAYER_PTR = nullptr;
-int16_t LAYER_UPDATE = 0;
-int16_t LAYER_UPDATE_X1 = INT16_MAX;
-int16_t LAYER_UPDATE_Y1 = INT16_MAX;
-int16_t LAYER_UPDATE_X2 = INT16_MIN;
-int16_t LAYER_UPDATE_Y2 = INT16_MIN;
-std::vector<LAYER_INFO> LAYERS;
-
-// BRUSH
-SDL_Texture* BRUSH_CURSOR_TEXTURE;
-SDL_Texture* BRUSH_TEXTURE;
-bool BRUSH_UPDATE = 0;
-
-int16_t BRUSH_X = 0;
-int16_t BRUSH_Y = 0;
-uint16_t BRUSH_W = 0;
-int16_t BRUSH_UPDATE_X1 = INT16_MAX;
-int16_t BRUSH_UPDATE_Y1 = INT16_MAX;
-int16_t BRUSH_UPDATE_X2 = INT16_MIN;
-int16_t BRUSH_UPDATE_Y2 = INT16_MIN;
-std::unique_ptr<COLOR[]> BRUSH_PIXELS;
-COLOR BRUSH_COLOR {255, 255, 255, 128};
-COLOR UNDO_COLOR{255, 0, 64, 192};
-COLOR* BRUSH_CURSOR_PIXELS;
-COLOR* BRUSH_CURSOR_PIXELS_CLEAR;
-SDL_Rect BRUSH_CURSOR_PIXELS_CLEAR_RECT;
-uint16_t BRUSH_CURSOR_PIXELS_CLEAR_POS;
-
-
-std::vector<std::unique_ptr<BRUSH_DATA>> BRUSH_LIST;
-uint16_t BRUSH_LIST_POS = 0;
-
-//SDL_Texture* CANVAS_TEXTURE;// = SDL_CreateTexture(RENDERER, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_STREAMING, CANVAS_W, CANVAS_H);
-//uint32_t* CANVAS_PIXELS;// = new uint32_t[CANVAS_W * CANVAS_H];
-
 bool MOUSEBUTTON_LEFT = false;
 bool MOUSEBUTTON_PRESSED_LEFT = false;
 bool MOUSEBUTTON_MIDDLE = false;
@@ -124,22 +52,6 @@ bool MOUSEBUTTON_PRESSED_RIGHT = false;
 // TOOL
 uint16_t CURRENT_TOOL = 0;
 
-COLOR COL_ACCENT{ 0xff, 0x00, 0x40, 0xc0 };
-COLOR COL_BGUPDATE{ 0x00, 0x00, 0x00, 0x01 };
-
-std::vector<UNDO_DATA> UNDO_LIST;
-uint16_t UNDO_POS = 0;
-uint16_t UNDO_UPDATE = 0;
-uint16_t UNDO_UPDATE_LAYER = 0;
-SDL_Rect UNDO_UPDATE_RECT = { 0, 0, 1, 1 };
-COLOR UNDO_COL = COL_ACCENT;
-
-// PRESET COLORS
-COLOR COL_WHITE{ 0xff, 0xff, 0xff, 0xff };
-COLOR COL_LTGRAY{ 0xc0, 0xc0, 0xc0, 0xff };
-COLOR COL_GRAY{ 0x80, 0x80, 0x80, 0xff };
-COLOR COL_DKGRAY{ 0x40, 0x40, 0x40, 0xff };
-COLOR COL_BLACK{ 0x00, 0x00, 0x00, 0xff };
 
 
 SDL_Cursor* create_system_cursor()
