+++ src/VARIABLES.h
@@ -54,107 +54,6 @@ extern int WINDOW_H;
 extern int DISPLAY_W;
 extern int DISPLAY_H;
 
-// CANVAS
-extern bool CANVAS_UPDATE;
-extern float CANVAS_ZOOM;
-extern float CANVAS_X;
-extern float CANVAS_Y;
-extern uint16_t CANVAS_W;
-extern uint16_t CANVAS_H;
-extern int16_t CANVAS_MOUSE_X;
-extern int16_t CANVAS_MOUSE_Y;
-extern int16_t CANVAS_MOUSE_CLICKX;
-extern int16_t CANVAS_MOUSE_CLICKY;
-extern int16_t CANVAS_MOUSE_CLICKPREVX;
-extern int16_t CANVAS_MOUSE_CLICKPREVY;
-extern int16_t CANVAS_MOUSE_PREVX;
-extern int16_t CANVAS_MOUSE_PREVY;
-extern int16_t CANVAS_MOUSE_CELL_X;
-extern int16_t CANVAS_MOUSE_CELL_Y;
-extern int16_t CANVAS_PITCH;
-
-extern SDL_Texture* BG_GRID_TEXTURE;
-extern int16_t CELL_W;
-extern int16_t CELL_H;
-extern int16_t BG_GRID_W;
-extern int16_t BG_GRID_H;
-
-extern float CANVAS_PANX;
-extern float CANVAS_PANY;
-
-extern float CANVAS_X_ANIM;
-extern float CANVAS_Y_ANIM;
-extern float CANVAS_W_ANIM;
-extern float CANVAS_H_ANIM;
-extern float CELL_W_ANIM;
-extern float CELL_H_ANIM;
-
-// LAYER
-extern uint16_t CURRENT_LAYER;
-extern COLOR* CURRENT_LAYER_PTR;
-extern int16_t LAYER_UPDATE;
-extern int16_t LAYER_UPDATE_X1;
-extern int16_t LAYER_UPDATE_Y1;
-extern int16_t LAYER_UPDATE_X2;
-extern int16_t LAYER_UPDATE_Y2;
-
-struct LAYER_INFO {
-	SDL_Texture* texture;
-	std::unique_ptr<COLOR[]> pixels;
-	int16_t x;
-	int16_t y;
-	uint8_t alpha;
-	SDL_BlendMode blendmode;
-};
-
-extern std::vector<LAYER_INFO> LAYERS;
-
-// BRUSH
-extern SDL_Texture* BRUSH_CURSOR_TEXTURE;
-extern SDL_Texture* BRUSH_TEXTURE;
-extern bool BRUSH_UPDATE;
-
-extern int16_t BRUSH_X;
-extern int16_t BRUSH_Y;
-extern uint16_t BRUSH_W;
-extern int16_t BRUSH_UPDATE_X1;
-extern int16_t BRUSH_UPDATE_Y1;
-extern int16_t BRUSH_UPDATE_X2;
-extern int16_t BRUSH_UPDATE_Y2;
-extern std::unique_ptr<COLOR[]> BRUSH_PIXELS;
-extern COLOR BRUSH_COLOR;
-extern COLOR UNDO_COLOR;
-extern COLOR* BRUSH_CURSOR_PIXELS;
-extern COLOR* BRUSH_CURSOR_PIXELS_CLEAR;
-extern SDL_Rect BRUSH_CURSOR_PIXELS_CLEAR_RECT;
-extern uint16_t BRUSH_CURSOR_PIXELS_CLEAR_POS;
-
-struct BRUSH_DATA
-{
-	int x = 0;
-	int y = 0;
-	int w = 0;
-	std::vector<uint8_t> alpha;
-
-	BRUSH_DATA(int _w) : alpha(_w* _w)
-	{
-		this->w = _w;
-		this->x = (-_w / 2);
-		this->y = (-_w / 2);
-	}
-
-	void set(int p, uint8_t a)
-	{
-		alpha[p] = a;
-	}
-};
-
-extern std::vector<std::unique_ptr<BRUSH_DATA>> BRUSH_LIST;
-extern uint16_t BRUSH_LIST_POS;
-
-//SDL_Texture* CANVAS_TEXTURE;// = SDL_CreateTexture(RENDERER, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_STREAMING, CANVAS_W, CANVAS_H);
-//uint32_t* CANVAS_PIXELS;// = new uint32_t[CANVAS_W * CANVAS_H];
-
 extern bool MOUSEBUTTON_LEFT;
 extern bool MOUSEBUTTON_PRESSED_LEFT;
 extern bool MOUSEBUTTON_MIDDLE;
@@ -165,49 +64,6 @@ extern bool MOUSEBUTTON_PRESSED_RIGHT;
 // TOOL
 extern uint16_t CURRENT_TOOL;
 
-// UNDO
-struct UNDO_DATA
-{
-	uint16_t layer = 0;
-	uint16_t x = 0;
-	uint16_t y = 0;
-	uint16_t w;
-	uint16_t h;
-	std::vector<COLOR> undo_pixels;
-	std::vector<COLOR> redo_pixels;
-
-	UNDO_DATA(uint16_t _w, uint16_t _h)
-		: w {_w}
-		, h {_h}
-		, undo_pixels(_w*_h)
-		, redo_pixels(_w*_h)
-	{
-	}
-
-	void set(uint16_t xx, uint16_t yy, COLOR prev_col, COLOR new_col)
-	{
-		int index = xx + yy * this->w;
-		undo_pixels[index] = prev_col;
-		redo_pixels[index] = new_col;
-	}
-};
-
-extern COLOR COL_ACCENT;
-extern COLOR COL_BGUPDATE;
-
-extern std::vector<UNDO_DATA> UNDO_LIST;
-extern uint16_t UNDO_POS;
-extern uint16_t UNDO_UPDATE;
-extern uint16_t UNDO_UPDATE_LAYER;
-extern SDL_Rect UNDO_UPDATE_RECT;
-extern COLOR UNDO_COL;
-
-// PRESET COLORS
-extern COLOR COL_WHITE;
-extern COLOR COL_LTGRAY;
-extern COLOR COL_GRAY;
-extern COLOR COL_DKGRAY;
-extern COLOR COL_BLACK;
 
 SDL_Cursor* create_system_cursor();
